requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0.3
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-extensions==3.2.3
django-jazzmin==2.6.0
django-rest-knox==4.2.0
djangorestframework==3.15.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.2.0
pycparser==2.21
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
text-unidecode==1.3
uritemplate==4.1.1
